requirements.txt
beautifulsoup4==4.4.1
certifi==2015.4.28
html5lib==0.9999999
limbo==5.0.3
pkg-resources==0.0.0
pyfiglet==0.7.4
PyYAML==3.12
requests==2.12.4
six==1.10.0
slackbot==0.4.1
slacker==0.9.30
webencodings==0.5
websocket-client==0.35.0
